Vial Keymap
There is a bug in the hex file affecting the initial Nice69 keymap. Otherwise, this bug does nothing to interfere with the rest of the board's operations. 

All Nice69 Elite-C's are shipping with this corrected keymap; but if there is a need to re-flash (or just flash) your Elite-C, you can use sideload this keymap into Vial for a quick fix!